fix convert and add UnitTest

+++ app/src/main/java/com/example/challengebtg/viewModel/CurrencyViewlModel.kt
@@ -80,7 +80,7 @@ class CurrencyViewlModel : ViewModel(), CoroutineScope {
             selectedCurrencyFrom ?: Currency(),
             selectedCurrencyTo ?: Currency(),
             quotes,
-            amount
+            amount.toDouble()
         )
 
         val resultString =
